make sidecar tests run everywhere.
