+++ pkg/sidecar/instance.go
@@ -1,5 +1,3 @@
-//+build linux
-
 package sidecar
 
 import (
