+++ pkg/sidecar/mock.go
@@ -1,5 +1,3 @@
-//+build linux
-
 package sidecar
 
 import (
